fix: add custom isMounted hook
