+++ client/src/app/routes/stardust/Block.tsx
@@ -3,7 +3,7 @@ import {
     MILESTONE_PAYLOAD_TYPE, TRANSACTION_PAYLOAD_TYPE,
     TAGGED_DATA_PAYLOAD_TYPE, milestoneIdFromMilestonePayload, IMilestonePayload
 } from "@iota/iota.js-stardust";
-import React, { useContext, useEffect, useRef, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { Link, RouteComponentProps } from "react-router-dom";
 import mainHeaderMessage from "../../../assets/modals/stardust/block/main-header.json";
 import metadataInfo from "../../../assets/modals/stardust/block/metadata.json";
@@ -12,6 +12,7 @@ import referencedBlocksInfo from "../../../assets/modals/stardust/block/mileston
 import taggedDataPayloadInfo from "../../../assets/modals/stardust/block/tagged-data-payload.json";
 import transactionPayloadInfo from "../../../assets/modals/stardust/block/transaction-payload.json";
 import { ServiceFactory } from "../../../factories/serviceFactory";
+import { useIsMounted } from "../../../helpers/hooks/useIsMounted";
 import { isMarketedNetwork } from "../../../helpers/networkHelper";
 import PromiseMonitor, { PromiseStatus } from "../../../helpers/promise/promiseMonitor";
 import { NameHelper } from "../../../helpers/stardust/nameHelper";
@@ -39,7 +40,7 @@ import "./Block.scss";
 const Block: React.FC<RouteComponentProps<BlockProps>> = (
     { history, match: { params: { network, blockId } } }
 ) => {
-    const isMounted = useRef(false);
+    const isMounted = useIsMounted();
     const { tokenInfo, bech32Hrp, protocolVersion } = useContext(NetworkContext);
     const [tangleCacheService] = useState(
         ServiceFactory.get<StardustTangleCacheService>(`tangle-cache-${STARDUST}`)
@@ -54,22 +55,16 @@ const Block: React.FC<RouteComponentProps<BlockProps>> = (
     >();
 
     useEffect(() => {
-        isMounted.current = true;
-
-        return () => {
-            isMounted.current = false;
-            if (updateMetadataTimerId) {
-                clearTimeout(updateMetadataTimerId);
-            }
-        };
-    }, []);
-
-    useEffect(() => {
         setBlockData({});
         setBlockMetadata({ blockTangleStatus: "pending" });
         // eslint-disable-next-line no-void
         void loadBlock(blockId);
-    }, [blockId]);
+        return () => {
+            if (updateMetadataTimerId) {
+                clearTimeout(updateMetadataTimerId);
+            }
+        };
+    }, [isMounted, blockId]);
 
     useEffect(() => {
         if (!blockData.blockError) {
