+++ client/src/app/routes/stardust/Foundry.tsx
@@ -1,9 +1,10 @@
 import { IAliasAddress, IFoundryOutput, IImmutableAliasUnlockCondition } from "@iota/iota.js-stardust";
 import { optional } from "@ruffy/ts-optional";
-import React, { useContext, useEffect, useRef, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { RouteComponentProps } from "react-router";
 import { Link } from "react-router-dom";
 import { ServiceFactory } from "../../../factories/serviceFactory";
+import { useIsMounted } from "../../../helpers/hooks/useIsMounted";
 import { isMarketedNetwork } from "../../../helpers/networkHelper";
 import PromiseMonitor, { PromiseStatus } from "../../../helpers/promise/promiseMonitor";
 import { formatAmount } from "../../../helpers/stardust/valueFormatHelper";
@@ -23,7 +24,7 @@ import "./Foundry.scss";
 const Foundry: React.FC<RouteComponentProps<FoundryProps>> = (
     { match: { params: { network, foundryId } } }
 ) => {
-    const isMounted = useRef(false);
+    const isMounted = useIsMounted();
     const { tokenInfo } = useContext(NetworkContext);
     const [tangleCacheService] = useState(
         ServiceFactory.get<StardustTangleCacheService>(`tangle-cache-${STARDUST}`)
@@ -37,7 +38,6 @@ const Foundry: React.FC<RouteComponentProps<FoundryProps>> = (
     const [controllerAlias, setControllerAlias] = useState<string>();
 
     useEffect(() => {
-        isMounted.current = true;
         const foundryLoadMonitor = new PromiseMonitor(status => {
             setJobToStatus(jobToStatus.set("loadFoundryDetails", status));
         });
@@ -66,10 +66,7 @@ const Foundry: React.FC<RouteComponentProps<FoundryProps>> = (
 
         // eslint-disable-next-line no-void
         void loadFoundryDetails();
-        return () => {
-            isMounted.current = false;
-        };
-    }, []);
+    }, [isMounted]);
 
 
     if (foundryError) {
