+++ client/src/helpers/hooks/useIsMounted.ts
@@ -1,0 +1,15 @@
+import { useEffect, useRef } from "react";
+
+export const useIsMounted = () => {
+    const isMounted = useRef(false);
+
+    useEffect(() => {
+      isMounted.current = true;
+      return () => {
+        console.log("unmounting");
+        isMounted.current = false;
+      };
+    }, []);
+
+    return isMounted;
+};
